+++ lib/awesome_bot.rb
@@ -1,5 +1,5 @@
 require 'awesome_bot/cli'
-require 'awesome_bot/command'
+require 'awesome_bot/check'
 
 # Validate awesome project URLs
 module AwesomeBot
