+++ lib/awesome_bot/check.rb
